pyproject.toml
[project]
name = "kocem"
version = "1.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "absl-py==2.1.0",
    "accelerate==1.0.1",
    "aiohappyeyeballs==2.4.3",
    "aiohttp==3.10.10",
    "aiosignal==1.3.1",
    "annotated-types==0.7.0",
    "anyio==4.6.2.post1",
    "attrs==24.2.0",
    "certifi==2024.8.30",
    "cfgv==3.4.0",
    "chardet==5.2.0",
    "charset-normalizer==3.4.0",
    "click==8.1.7",
    "colorama==0.4.6",
    "contourpy==1.3.0",
    "coverage==7.6.4",
    "cycler==0.12.1",
    "dataproperty==1.0.1",
    "datasets==3.1.0",
    "dill==0.3.8",
    "distlib==0.3.9",
    "distro==1.9.0",
    "et-xmlfile==1.1.0",
    "evaluate==0.4.3",
    "execnet==2.1.1",
    "fastapi>=0.116.1",
    "filelock==3.16.1",
    "fire==0.7.0",
    "fonttools==4.54.1",
    "frozenlist==1.5.0",
    "fsspec==2024.9.0",
    "h11==0.14.0",
    "httpcore==1.0.6",
    "httpx==0.27.2",
    "huggingface-hub==0.26.1",
    "identify==2.6.1",
    "idna==3.10",
    "iniconfig==2.0.0",
    "jinja2==3.1.4",
    "jiter==0.6.1",
    "joblib==1.4.2",
    "jsonlines==4.0.0",
    "kiwisolver==1.4.7",
    "langchain>=0.3.27",
    "langchain-anthropic>=0.3.19",
    "langchain-community>=0.3.27",
    "langchain-core>=0.3.74",
    "langchain-openai>=0.2.5",
    "langserve>=0.3.1",
    "lxml==5.3.0",
    "markupsafe==3.0.2",
    "matplotlib==3.9.2",
    "mbstrdecoder==1.1.3",
    "more-itertools==10.5.0",
    "mpmath==1.3.0",
    "multidict==6.1.0",
    "multiprocess==0.70.16",
    "mypy==1.13.0",
    "mypy-extensions==1.0.0",
    "networkx==3.4.2",
    "nltk==3.9.1",
    "nodeenv==1.9.1",
    "numexpr==2.10.1",
    "numpy==2.1.2",
    "openai==1.52.2",
    "openpyxl==3.1.5",
    "packaging==24.1",
    "pandas==2.2.3",
    "pathvalidate==3.2.1",
    "peft==0.13.2",
    "pillow==11.0.0",
    "platformdirs==4.3.6",
    "pluggy==1.5.0",
    "portalocker==2.10.1",
    "pre-commit==4.0.1",
    "propcache==0.2.0",
    "psutil==6.1.0",
    "pyarrow==17.0.0",
    "pybind11==2.13.6",
    "pydantic==2.9.2",
    "pydantic-core==2.23.4",
    "pyparsing==3.2.0",
    "pytablewriter==1.2.0",
    "pytest==8.3.3",
    "pytest-cov==5.0.0",
    "pytest-xdist==3.6.1",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.0.1",
    "pytz==2024.2",
    "pywin32==308",
    "pyyaml==6.0.2",
    "regex==2024.9.11",
    "requests==2.32.3",
    "rouge-score==0.1.2",
    "sacrebleu==2.4.3",
    "safetensors==0.4.5",
    "scikit-learn==1.5.2",
    "scipy==1.14.1",
    "seaborn==0.13.2",
    "setuptools==75.1.0",
    "six==1.16.0",
    "sniffio==1.3.1",
    "sqlitedict==2.1.0",
    "sympy==1.13.1",
    "tabledata==1.3.3",
    "tabulate==0.9.0",
    "tcolorpy==0.1.6",
    "tenacity==9.0.0",
    "termcolor==2.5.0",
    "threadpoolctl==3.5.0",
    "tiktoken==0.8.0",
    "tokenizers==0.20.1",
    "torch==2.5.0",
    "tqdm==4.66.5",
    "tqdm-multiprocess==0.0.11",
    "transformers==4.46.0",
    "typepy==1.3.2",
    "typing-extensions==4.12.2",
    "tzdata==2024.2",
    "urllib3==2.2.3",
    "virtualenv==20.27.0",
    "wheel==0.44.0",
    "word2number==1.1",
    "xxhash==3.5.0",
    "yarl==1.16.0",
    "zstandard==0.23.0",
]
